requirements.txt
Django==6.0.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
Pillow

# Code Quality
ruff==0.1.5
black==23.9.1
